requirements.txt
# FX Daily Bias Simplified - Requirements
# Version 1.0.0

# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
requests>=2.28.0

# Testing dependencies
pytest>=7.0.0
pytest-benchmark>=4.0.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Add parquet support
pyarrow>=10.0.0